+++ packages/terra-table/CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## Unreleased
 
+* Added
+  * Added the ability to select a cell by pressing Escape on an auto focused button or anchor element.
+
 * Fixed
   * Fixed an issue where the click event for buttons inside table cells was not triggered by the spacebar.
-  
+
 ## 5.15.0 - (March 21, 2024)
 
 * Changed